requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
pandas==2.1.1
transformers==4.38.2
torch==2.6.0
sentencepiece==0.2.0
requests==2.32.3
streamlit-webrtc==0.47.1
numpy>=1.24.3
av>=9.2.0 